Compress 8bit

+++ src/array_prototype/dnn/dnn_descriptor_runner.ts
@@ -1,7 +1,7 @@
 namespace WebDNN {
     export interface DNNDescriptorRunner {
         compile(): Promise<void>;
-        loadWeights(weightsData: Float32Array): Promise<void>;
+        loadWeights(weightsData: Uint8Array): Promise<void>;
         run(): Promise<void>;
         getInputViews(): Promise<Float32Array[]>;
         getOutputViews(): Promise<Float32Array[]>;
